fix: følger go convention på pakker og filnavn
