+++ cmd/devicecmd/configcmd.go
@@ -1,4 +1,4 @@
-package deviceCmd
+package devicecmd
 
 import (
 	"fmt"
