+++ cmd/postgrescmd/preparecmd.go
@@ -1,4 +1,4 @@
-package postgresCmd
+package postgrescmd
 
 import (
 	"bufio"
@@ -11,16 +11,22 @@ import (
 	"github.com/urfave/cli/v2"
 )
 
-func revokeCommand() *cli.Command {
+func prepareCommand() *cli.Command {
 	return &cli.Command{
-		Name:  "revoke",
-		Usage: "Revoke access to your postgres instance for the role 'cloudsqliamuser'",
-		Description: `Revoke will revoke the role 'cloudsqliamuser' access to the
-tables in the postgres instance. This is done by connecting using the application
-credentials and modify the permissions on the public schema.
+		Name:  "prepare",
+		Usage: "Prepare your postgres instance for use with personal accounts",
+		Description: `Prepare will prepare the postgres instance by connecting using the
+application credentials and modify the permissions on the public schema.
+All IAM users in your GCP project will be able to connect to the instance.
 
 This operation is only required to run once for each postgresql instance.`,
 		ArgsUsage: "appname",
+		Flags: []cli.Flag{
+			&cli.BoolFlag{
+				Name:  "all-privs",
+				Usage: "Gives all privalges to users",
+			},
+		},
 		Before: func(context *cli.Context) error {
 			if context.Args().Len() != 1 {
 				return fmt.Errorf("missing name of app")
@@ -31,6 +37,7 @@ This operation is only required to run once for each postgresql instance.`,
 		Action: func(context *cli.Context) error {
 			appName := context.Args().First()
 
+			allPrivs := context.Bool("all-privs")
 			namespace := context.String("namespace")
 			cluster := context.String("context")
 			database := context.String("database")
@@ -44,7 +51,7 @@ This operation is only required to run once for each postgresql instance.`,
 				return fmt.Errorf("cancelled by user")
 			}
 
-			return postgres.RevokeAccess(context.Context, appName, namespace, cluster, database)
+			return postgres.PrepareAccess(context.Context, appName, namespace, cluster, database, allPrivs)
 		},
 	}
 }
