+++ cmd/postgrescmd/psqlcmd.go
@@ -1,4 +1,4 @@
-package postgresCmd
+package postgrescmd
 
 import (
 	"fmt"
@@ -6,24 +6,16 @@ import (
 	"github.com/urfave/cli/v2"
 )
 
-func proxyCommand() *cli.Command {
+func psqlCommand() *cli.Command {
 	return &cli.Command{
-		Name:        "proxy",
-		Usage:       "Create a proxy to a Postgres instance",
-		Description: "Update IAM policies by giving your user the a timed sql.cloudsql.instanceUser role, then start a proxy to the instance.",
+		Name:        "psql",
+		Usage:       "Connect to the database using psql",
+		Description: "Create a shell to the postgres instance by opening a proxy on a random port (see the proxy command for more info) and opening a psql shell.",
 		ArgsUsage:   "appname",
 		Flags: []cli.Flag{
 			&cli.BoolFlag{
 				Name:    "verbose",
 				Aliases: []string{"v"},
-			},
-			&cli.UintFlag{
-				Name:  "port",
-				Value: 5432,
-			},
-			&cli.StringFlag{
-				Name:  "host",
-				Value: "localhost",
 			},
 		},
 		Before: func(context *cli.Context) error {
@@ -40,10 +32,8 @@ func proxyCommand() *cli.Command {
 			cluster := context.String("context")
 			database := context.String("database")
 			verbose := context.Bool("verbose")
-			port := context.Uint("port")
-			host := context.String("host")
 
-			return postgres.RunProxy(context.Context, appName, cluster, namespace, database, host, port, verbose)
+			return postgres.RunPSQL(context.Context, appName, cluster, namespace, database, verbose)
 		},
 	}
 }
